More testing
ImageParser passes testCubeStraightCrop, which is a version of
testCubeStraight that is perfectly square and has very little space
around the edges of the cube. These stipulations should be sufficient
but unnecessary for successful parsing, but as of now, both are
necessary. More testing is needed to discover why.

CubeParser returns an invalid cube (invalid centers) when applied to
Cube1. More testing is needed to discover whether this is the fault of
ImageParser or CubeParser. Suggesting: make 6 images similar to
testCubeStraightCrop (idealized, hard edges, sharp contrast images)
which form a valid cube. Make sure ImageParser can handle these (it
should) and then try them with CubeParser.